+++ include/helpers.h
@@ -1,4 +1,9 @@
+#ifndef HELPERS_H_
+#define HELPERS_H_
+
 void SetError();
 void ClearError();
 
 void EndienSwap(void *data, uint8_t n);
+
+#endif
