+++ platformio.ini
@@ -14,4 +14,4 @@ board = uno
 framework = arduino
 upload_protocol = arduino
 build_src_filter = +<**.c*> +<**.h> -<.git/>
-build_flags = -D __AVR_ATmega328P__ -Wl,-u,vfprintf -lprintf_flt -lm
+build_flags = -D __AVR_ATmega328P__ -Wl,-u,vfprintf -lprintf_flt -lmran
