+++ src/main.cpp
@@ -12,11 +12,6 @@ PIDController pidcontroller;
 float Kp = 1.0f, Ki = 0.00f, Kd = 0.0f;
 float Hz = 10.0f;
 
-void SetError();
-void ClearError();
-void PID_Start();
-void PID_Stop();
-
 int main()
 {
 	sei();
@@ -37,10 +32,6 @@ int main()
 	uint8_t msg_size = sprintf(msg, "min %u max %u\n", IR_min, IR_max);
 	Bluetooth_Send(msg, msg_size + 1);
 
-	// init PID timer
-	TCCR1B = 0b11 << WGM12 | 0b100 << CS10;
-	ICR1 = 15625;
-
 	// start_motors();
 	// drive_motors(0.5f, 0.0f);
 	PID_Start();
@@ -49,16 +40,6 @@ int main()
 		;
 
 	return 0;
-}
-
-void PID_Start()
-{
-	TIMSK1 |= 0b1 << OCIE1A;
-}
-
-void PID_Stop()
-{
-	TIMSK1 &= ~(0b1 << OCIE1A);
 }
 
 RunningDiagnostics diagnostics;
